clear previous testing framework
