+++ runtime/integration-tests/src/generic/mod.rs
@@ -1,8 +1,3 @@
-// For generic modules we can reactive the unused warn disabled on lib.rs
-#![warn(unused)]
-// Allow dead code for utilities not used yet
-#![allow(dead_code)]
-
 pub mod env;
 pub mod envs {
 	pub mod fudge_env;
