+++ runtime/integration-tests/src/utils/mod.rs
@@ -10,29 +10,5 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-use cfg_primitives::Balance;
-use cfg_types::tokens::CurrencyId;
-
 pub mod accounts;
-pub mod env;
-pub mod evm;
-pub mod extrinsics;
-pub mod genesis;
 pub mod logs;
-pub mod time;
-pub mod tokens;
-
-/// The relay native token's asset id
-pub const RELAY_ASSET_ID: CurrencyId = CurrencyId::ForeignAsset(5);
-/// The Glimmer asset id
-pub const GLMR_CURRENCY_ID: CurrencyId = CurrencyId::ForeignAsset(4);
-/// The AUSD asset id
-pub const AUSD_CURRENCY_ID: CurrencyId = CurrencyId::ForeignAsset(3);
-/// The USDT asset id
-pub const USDT_CURRENCY_ID: CurrencyId = CurrencyId::ForeignAsset(1);
-/// The EVM Chain id of Moonbeam
-pub const MOONBEAM_EVM_CHAIN_ID: u64 = 1284;
-
-pub const GLMR_ED: Balance = 1_000_000;
-pub const AUSD_ED: Balance = 1_000_000_000;
-pub const USDT_ED: Balance = 10_000;
